2nd fix clang-tidy and clang-format

+++ tasks/seq/shishkarev_a_gift_wraping_algorithm/func_tests/main.cpp
@@ -2,6 +2,7 @@
 
 #include <cstddef>
 #include <cstdint>
+#include <cstdlib>
 #include <fstream>
 #include <memory>
 #include <string>
@@ -18,7 +19,7 @@ TEST(shishkarev_a_gift_wraping_algorithm_seq, test_convex_hull_simple) {
   std::vector<shishkarev_a_gift_wraping_algorithm_seq::Vertex> out(kCount);
 
   for (size_t i = 0; i < kCount; i++) {
-    in[i] = {.x = (static_cast<int>(rand() % 100)), .y = (static_cast<int>(rand() % 100))};
+    in[i] = {.x = rand() % 100, .y = rand() % 100};
   }
 
   auto task_data_seq = std::make_shared<ppc::core::TaskData>();
@@ -50,7 +51,7 @@ TEST(shishkarev_a_gift_wraping_algorithm_seq, test_convex_hull_from_file) {
   std::vector<shishkarev_a_gift_wraping_algorithm_seq::Vertex> out(count);
 
   for (size_t i = 0; i < count; i++) {
-    in[i] = {.x = (static_cast<int>(rand() % 100)), .y = (static_cast<int>(rand() % 100))};
+    in[i] = {.x = rand() % 100, .y = rand() % 100};
   }
 
   auto task_data_seq = std::make_shared<ppc::core::TaskData>();
